split //daml-lf/valiation/upgrade-tests  this should reduce flakyness and running time
