+++ sdk/daml-lf/validation/src/test/scala/com/daml/lf/validation/upgrade/UpgradesSpecLedgerAPIWithValidation.scala
@@ -1,0 +1,9 @@
+// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.lf.validation
+package upgrade
+
+class UpgradesSpecLedgerAPIWithValidation
+    extends UpgradesSpecLedgerAPI("Ledger API with validation")
+    with LongTests
